ct week4 done #4
