+++ lecture/ch4/ch4/1-A_permutation.cpp
