+++ lecture/ch4/ch4/1-D.cpp
@@ -1,0 +1,24 @@
+//// 백준 10988 팰린드롬인지 확인하기 -> 앞, 뒤가 똑같은 문자열
+//#include <bits/stdc++.h>
+//using namespace std;
+//
+//string s, temp;
+//
+//int main() {
+//
+//	ios_base::sync_with_stdio(false);
+//	cin.tie(NULL);
+//	cout.tie(NULL);
+//
+//	cin >> s;
+//	temp = s;
+//
+//	reverse(temp.begin(), temp.end());
+//	if (temp == s) {
+//		cout << 1 << "\n";
+//	}
+//	else {
+//		cout << 0 << "\n";
+//	}
+//	return 0;
+//}
